backend/requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# LangChain and AI
langchain>=0.1.0
langchain-cohere>=0.1.0
langchain-community>=0.0.10
cohere>=5.1.4

# Database drivers
pymongo>=4.6.0
mysql-connector-python>=8.2.0
aiomysql>=0.2.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Configuration and utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.25.0
pydantic>=2.5.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database async support
motor>=3.3.0
aiofiles>=23.2.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Background tasks
redis>=5.0.0
celery>=5.3.0

# Vector database and embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0
